Setup notifications for events from task manager #124 * Refactoring
