+++ application-task-default/src/main/java/com/xwiki/task/internal/notifications/taskchanged/TaskChangedEventFactory.java
@@ -18,7 +18,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package com.xwiki.task.internal.notifications;
+package com.xwiki.task.internal.notifications.taskchanged;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -48,13 +48,6 @@ public class TaskChangedEventFactory
 {
     private static final LocalDocumentReference TASK_CLASS =
         new LocalDocumentReference("TaskManager", "TaskManagerClass");
-
-    /**
-     * Constructor. This is essentially a static class.
-     */
-    public TaskChangedEventFactory()
-    {
-    }
 
     /**
      * Get a list of events generated by the last update of the document, one for each specified field.
@@ -87,7 +80,7 @@ public class TaskChangedEventFactory
      * @param propertyName the property to retrieve
      * @return the value of the desired property, or null if the property doesn't exist or has unsupported type.
      */
-    public Optional<Object> getPropertyValue(BaseObject obj, String propertyName)
+    private Optional<Object> getPropertyValue(BaseObject obj, String propertyName)
     {
         if (null == obj) {
             return Optional.empty();
@@ -115,7 +108,7 @@ public class TaskChangedEventFactory
      * @param taskPage the parent page of the modified object
      * @return the event describing the changes done to the specified field, null when no change occurred
      */
-    public Optional<TaskChangedEvent> getFieldChangedEvent(BaseObject currentObject, BaseObject previousObject,
+    private Optional<TaskChangedEvent> getFieldChangedEvent(BaseObject currentObject, BaseObject previousObject,
         String propertyName, XWikiDocument taskPage)
     {
         Object currentValue = getPropertyValue(currentObject, propertyName).orElse(null);
