+++ application-task-default/src/test/java/com/xwiki/task/notifications/taskchanged/TaskChangedEventListenerTest.java
@@ -19,7 +19,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package com.xwiki.task.notifications;
+package com.xwiki.task.notifications.taskchanged;
 
 import java.util.stream.Stream;
 
@@ -43,9 +43,9 @@ import com.xpn.xwiki.XWikiContext;
 import com.xpn.xwiki.doc.XWikiDocument;
 import com.xpn.xwiki.internal.event.XObjectAddedEvent;
 import com.xpn.xwiki.internal.event.XObjectUpdatedEvent;
-import com.xwiki.task.internal.notifications.TaskChangedEvent;
-import com.xwiki.task.internal.notifications.TaskChangedEventListener;
-import com.xwiki.task.internal.notifications.TaskChangedEventNotificationListener;
+import com.xwiki.task.internal.notifications.taskchanged.TaskChangedEvent;
+import com.xwiki.task.internal.notifications.taskchanged.TaskChangedEventListener;
+import com.xwiki.task.internal.notifications.taskchanged.TaskChangedEventNotificationListener;
 import com.xwiki.task.model.Task;
 
 import static org.mockito.ArgumentMatchers.any;
